+++ KaiFighter/KaiFighterGame/Interfaces/IKiller.cs
@@ -5,6 +5,6 @@
         /// <summary>
         /// Determines how much damage can the object deal
         /// </summary>
-        int Damage { get; set; }
+        double Damage { get; set; }
     }
 }
